==> workflow/03_generate_figures.Rmd <==
---
title: "03_generate_figures"
author: "Leeya Pressburger"
date: "`r Sys.Date()`"
output:
  bookdown::html_document2:
    fig_caption: yes
    number_sections: yes
    toc: true
    toc_depth: 2
    toc_float: true
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
```

### R setup

```{r intro, warning=FALSE, message=FALSE}
# Load necessary packages, read in workflow settings
library(hector)
library(relaimpo)
library(dplyr)
library(tidyr)
library(ggplot2)
library(gridExtra)
library(ggExtra)
library(GGally)
library(bookdown)
library(data.table) 

theme_set(theme_light())

source("common_data.R")

# Read in data files, generated by 02_pre_processing.Rmd
runlist <- fread("../output/runlist.csv")
trk_output_final <- fread("../output/trk_output_final.csv")

```

### Results

## Sources of atmospheric CO~2~

### Sources by pool

``` {r atmosphere-by-source, message = FALSE, fig.cap = paste0(" Fraction of CO2 in the atmosphere by source pool in ", MAIN_SCENARIO, ". Lines show median; shaded areas show ±1 s.d. of median (darker) and minimum and maximum of ensemble (lighter).")}
# This chunk looks at sources of CO2 in the atmosphere by pool

# Isolate the atmosphere pool to better understand its composition 
atmosphere_pool <- trk_output_final %>%
  filter(pool_name == "atmos_co2")

# This plot looks at source fraction by source pool in the atmosphere 
atm_summary <- atmosphere_pool %>%
  filter(ssp == MAIN_SCENARIO) %>%
  group_by(year, source_name) %>%
  summarize(sf_sd = sd(source_fraction), 
            sf_min = min(source_fraction),
            sf_max = max(source_fraction),
            sf_median = median(source_fraction))

# Order pools by contributions to atmospheric CO2
atm_summary$source_name <- factor(atm_summary$source_name, 
                                  levels = rev(c("HL", "LL", "detritus_c",
                                                 "intermediate", "deep", 
                                                 "veg_c", "soil_c", "earth_c", 
                                                 "atmos_co2")))

# Plot
atm_plot <- atm_summary %>%
  filter(year > 1800) %>%
  ggplot(aes(year)) +
  geom_ribbon(aes(ymin = sf_min, ymax = sf_max), alpha = 0.2, fill = "#01665e") +
  geom_ribbon(aes(ymin = sf_median - sf_sd, ymax = sf_median + sf_sd), alpha = 0.2, fill = "#01665e") +
  geom_line(aes(y = sf_median), color = "#01665e") +
  facet_wrap(~source_name, scales = "free",
             labeller = labeller(source_name = pool_labels)) +
  labs(x = "Year",
       y = expression(Source~fraction~of~CO[2]~"in"~atmosphere)) ; atm_plot

# Compute the mean and median percentage of anthropogenic CO2 in the 
# atmosphere for key years
percent_summary <- atmosphere_pool %>%
  filter(year %in% c(2015, 2050, 2100, 2300),
         source_name == "earth_c") %>%
  group_by(ssp, year) %>%
  summarize(avg = mean(source_fraction),
            med = median(source_fraction))

# Write output files
write.csv(percent_summary, "../output/percent_anthro_co2_in_atmos.csv")

ggsave(plot = atm_plot, "../figures/figure3.png", height = 6, width = 9, units = "in")

```

### Anthropogenic emissions in the atmosphere across scenarios

``` {r earthc-in-atmos, fig.cap = paste0(" Fraction of atmospheric CO2 from anthropogenic emissions over time across scenarios. Line shows median; shaded areas show ±1 s.d. of median (darker) and minimum and maximum of ensemble (lighter).")}
# This chunk looks at the atmosphere pool with just one source (earth_c) 
# and plots source fraction over time with a confidence interval for all SSPs

ff <- atmosphere_pool %>% 
  filter(source_name == "earth_c") %>%
  group_by(ssp, year) %>%
  mutate(sf_sd = sd(source_fraction), 
         sf_min = min(source_fraction),
         sf_max = max(source_fraction),
         sf_median = median(source_fraction))

rm(atmosphere_pool)

# Plot
atm_fraction_plot <- ggplot(ff, aes(year,  fill = ssp)) +
  geom_line(aes(y = sf_median, color = ssp), size = 1) +
  geom_ribbon(aes(ymin = sf_median - sf_sd, ymax = sf_median + sf_sd), alpha = 0.5) +
  scale_color_manual(breaks = names(ssp_labels),
                     labels = ssp_labels,
                     values = c("#8c510a", "#dfc27d", "#c7eae5", "#01665e")) +
  scale_fill_manual(breaks = names(ssp_labels),
                    labels = ssp_labels,
                    values = c("#8c510a", "#dfc27d", "#c7eae5", "#01665e")) +
  labs(x = "Year",
       y = expression(Fraction~of~atmospheric~CO[2]~from~anthropogenic~emissions),
       color = "Scenario",
       fill = "Scenario") +
  scale_x_continuous(expand = c(0, 0)) ; atm_fraction_plot

# Save figure
ggsave(plot = atm_fraction_plot, "../figures/figure4.png", height = 6, width = 9, units = "in")

```

``` {r cleanup-source, echo = FALSE}
rm(list = c("ff", "atm_summary", "percent_summary",
            "atm_plot", "atm_fraction_plot"))
gc()
```

## Destination of anthropogenic emissions

### Destination fraction

```{r destination, warning = FALSE, fig.cap = paste0(" Final destination pools of anthropogenic emissions averaged over ", SSP_runs[MAIN_SCENARIO], " runs in ", MAIN_SCENARIO, ".")}
# This chunk looks at where CO2 ends up in the atmosphere by computing the
# "destination fraction" of anthropogenic CO2 in all other pools

# Step 1: compute source_quantity = pool_value * source_fraction
# Step 2: for each year and source_name, compute destination_fraction

dest_output <- trk_output_final %>%
  filter(source_name == "earth_c", pool_name != "earth_c") %>%
  mutate(source_quantity = pool_value * source_fraction) %>%
  group_by(year, run_number) %>%
  summarize(destination_fraction = source_quantity / sum(source_quantity),
            pool_name = pool_name,
            .groups = "drop") %>% 
  ungroup()

dest_output$ssp <- runlist$ssp[dest_output$run_number]

# Compute the mean destination fraction
dest_average <- dest_output %>%
  group_by(ssp, year, pool_name) %>%
  mutate(dest_avg = mean(destination_fraction),
         pool_name = factor(pool_name, 
                            levels = rev(c("deep", "intermediate",
                                           "LL", "HL", "detritus_c", 
                                           "veg_c", "soil_c", "atmos_co2")))) %>%
  filter(year > 1800) %>%
  ungroup()

rm(dest_output)
gc()

# Looking at average pool destination
average_summary <- dest_average %>% 
  filter(year == 2300) %>% 
  mutate(pool_name = gsub("atmos_co2", "atmosphere", pool_name),
         pool_name = gsub("soil_c", "land", pool_name),
         pool_name = gsub("veg_c", "land", pool_name),
         pool_name = gsub("detritus_c", "land", pool_name),
         pool_name = gsub("HL", "ocean", pool_name),
         pool_name = gsub("LL", "ocean", pool_name),
         pool_name = gsub("deep", "ocean", pool_name),
         pool_name = gsub("intermediate", "ocean", pool_name)) %>%
  group_by(run_number, pool_name) %>%
  summarize(ssp = ssp,
            sum = sum(dest_avg)) %>%
  distinct() %>%
  ungroup() %>%
  group_by(ssp, pool_name) %>%
  summarize(avg = mean(sum)) %>%
  pivot_wider(names_from = pool_name, values_from = avg)

# Write output file
write.csv(average_summary, "../output/average_summary_2300.csv")

# Plot
dest_plot <- dest_average %>%  
  group_by(ssp, year, pool_name) %>%
  filter(run_number == min(run_number)) %>%
  ggplot(aes(year, dest_avg, fill = pool_name)) +
  geom_area() +
  scale_fill_brewer(breaks = names(o_pool_labels),
                    labels = o_pool_labels,
                    palette = "BrBG",
                    direction = -1) +
  labs(x = "Year",
       y = "Fraction of anthropogenic emissions",
       fill = "Destination") +
  scale_x_continuous(expand = c(0, 0)) +
  facet_wrap(~ssp,
             labeller = labeller(ssp = ssp_labels)) +
  theme(axis.text.x = element_text(angle = 90)) ; dest_plot

ggsave(plot = dest_plot, "../figures/figure5.png", height = 6, width = 9, units = "in")

```

### What controls how much anthropogenic CO2 ends up in the atmosphere?

```{r relative-importance, message = FALSE, echo = FALSE, warning = FALSE, fig.cap = " Relative importance of parameters over time in controlling the atmosphere as a destination of anthropogenic emissions. Relative importance is the R^2 contribution averaged over orderings among regressors; here, we normalize the values so that they sum to one."}
# This chunk computes the "relative importance" of each Hector parameter
# on controlling how much anthropogenic CO2 ends up in the atmosphere

# Process data
# Filter to just atmosphere pool, and remove unneeded columns
dest <- dest_average %>% 
  left_join(runlist, by = c("run_number", "ssp")) %>%
  as.data.frame() %>% 
  filter(pool_name == "atmos_co2") %>%
  select(-c(run_number, pool_name, dest_avg))

rm(runlist)

# Function to calculate relative importance of parameters in
# explaining how much destination_fraction varies for a given year of data. 
# This uses relaimpo::calc.relimp()
# See https://www.r-bloggers.com/2012/08/the-relative-importance-of-predictors-let-the-games-begin/
# for a useful primer, along with the calc.reimp help page
calc_relimp <- function(x) {
  yr <- x$year[1] # record year...
  x$year <- NULL # and drop year column
  sce <- x$ssp[1]
  x$ssp <- NULL
  # Fit model. The formula notation means "destination fraction as a
  # function of all other terms" (which is why we dropped year)
  m <- lm(destination_fraction ~ ., data = x)
  # Calculate relative importance metrics and extract 'lmg' results
  # 'lmg' is the R^2 contribution averaged over orderings among regressors;
  # should sum to one because we're using relative importances (rela=TRUE)
  relimp <- try(calc.relimp(m, type = "lmg", rela = TRUE)@lmg)
  if(class(relimp) == "try-error") {
    message(yr, " calc.relimp error")
    return(NULL)
  }
  # Return a data frame: year, parameter, relative importance
  tibble(ssp = sce, year = yr, param = names(relimp), value = relimp)
  
}

# Scale ocean parameters down for relaimpo calculation
dest <- dest %>%
  mutate(TT = TT / 1e7,
         TU = TU / 1e7,
         TWI = TWI / 1e7,
         TID = TID / 1e7)

# Run calc_relimp() above on each year's data
dest_split <- split(dest, list(dest$year, dest$ssp))
relimp_out <- lapply(dest_split, FUN = calc_relimp) %>% 
  bind_rows()

# ...and plot!
importance_plot <- relimp_out %>%
  mutate(param = factor(param, levels = rev(c("NPP_FLUX0", "Q10_RH", "BETA", "ECS", 
                                              "DIFFUSIVITY", "AERO_SCALE",
                                              "LUC_SCALE", "TT", "TU", "TWI", "TID")))) %>%
  filter(year > 1999) %>%
  ggplot(aes(year, value, fill = param)) + 
  geom_area() +
  coord_cartesian(expand = FALSE) +
  scale_fill_brewer(palette = "BrBG",
                    breaks = c("TID", "TWI", "TU", "TT", "LUC_SCALE", "AERO_SCALE", "DIFFUSIVITY",
                               "ECS", "BETA", "Q10_RH","NPP_FLUX0"),
                    labels = c("Intermed-deep exchange", "Warm-intermed exchange", 
                               "High-lat overturning", "Thermohaline overturning", 
                               "LUC emissions", "Aerosol forcing",
                               "Ocean diffusivity", "ECS",
                               "CO2 fertilization", "Q10",
                               "Pre-industrial NPP")) +
  facet_wrap(~ssp,
             labeller = labeller(ssp = ssp_labels)) +
  labs(x = "Year",
       y = expression(Relative~importance~"for"~anthropogenic~CO[2]~"in"~the~atmosphere),
       fill = "Parameter") +
  theme(axis.text.x = element_text(angle = 90),
        panel.spacing=unit(1, "lines")) +
  scale_x_continuous(expand = c(0, 0)) ; importance_plot

ggsave(plot = importance_plot, "../figures/figure6.png", height = 6, width = 9, units = "in")

```

``` {r cleanup-destination, echo = FALSE}
rm(list = c("dest_plot", "all_years_tracking", "dest", "dest_minimal", 
            "dest_output", "dest_average", "average_summary",
            "relimp_out", "dest_split", "importance_plot"))

gc()
```

## Airborne fraction

### Airborne fraction and "fraction of emissions remaining in atmosphere"

```{r airborne, warning = FALSE, fig.cap = " Airborne fraction over time, by scenario. The yellow line and band shows the actual amount of anthropogenic emissions remaining in the atmosphere from the model's carbon-tracking mechanism, whereas the blue line shows the actual airborne fraction."}

# Prep the data
atm_earth_diffs <- trk_output_final %>%
  # Isolate atmosphere and earth total pool values
  filter(pool_name %in% c("atmos_co2", "earth_c")) %>% 
  group_by(run_number, year, pool_name, ssp) %>% 
  summarise(pool_value = mean(pool_value), .groups = "drop") %>% 
  arrange(year) %>%
  # Put atmosphere and earth in separate columns and compute diffs
  pivot_wider(names_from = "pool_name", values_from = "pool_value") %>% 
  group_by(run_number, ssp) %>%
  mutate(atm_diff = c(NA, diff(atmos_co2)), 
         earth_diff = c(NA, diff(earth_c)))

# Compute AF (classical definition of delta CO2atm / emissions) 
# using various window sizes
af_results <- list()
for(groupsize in seq(1, 40, length.out = 10)) {
  ngroups = length(unique(atm_earth_diffs$year)) / groupsize
  
  atm_earth_diffs %>% 
    group_by(run_number, ssp) %>% 
    mutate(group = cut(year, breaks = ngroups)) %>% 
    # compute differences over the group of years
    group_by(run_number, group, ssp) %>%
    summarise(year = max(year), 
              atm_diff = sum(atm_diff), 
              earth_diff = sum(earth_diff), 
              n = n(),
              af = atm_diff / -earth_diff,
              .groups = "drop") ->
    af_results[[as.character(groupsize)]]
}

af_results <- bind_rows(af_results, .id = "window_size") %>% 
  mutate(window_size = as.integer(window_size))

# Compute cumulative emissions over time
# We use the change in earth_c to compute emissions; note this will not work
# if using a scenario with fossil fuel sequestration (atmosphere -> earth)

# This function computes annual cumulative earth_c emissions 
# for a given tracking dataset
min_runs <- trk_output_final %>%
  group_by(ssp) %>%
  summarize(mrun = min(run_number))

calc_emissions <- function(data){
  data %>% 
    filter(pool_name == "earth_c",
           source_name == "earth_c",
           run_number %in% min_runs$mrun) %>%
    arrange(year) %>%
    group_by(ssp) %>%
    # the NA at the end here is so that cumulative emissions 'line up' correctly
    # with the source_quantity (earth-origin atmosphere C) calculated below
    mutate(emissions = -c(diff(pool_value), NA),
           # relative to core$trackingDate
           cumulative_emissions = cumsum(emissions)) %>%
    select(year, ssp, cumulative_emissions) %>%
    arrange(ssp, year)
}

# This function takes a given tracking dataset and its corresponding cumulative 
# emissions dataset as found above to calculate what fraction of the cumulative
# emissions come from earth_c emisions that are now in the atmosphere
calc_emissions_fraction <- function(data, emissions_data){
  data %>%
    filter(pool_name == "atmos_co2", 
           source_name == "earth_c") %>%
    # How much earth_c is in atmos_co2? 
    mutate(source_quantity = pool_value * source_fraction) %>%
    left_join(emissions_data, by = c("year", "ssp")) %>%
    select(run_number, ssp, year, source_quantity, cumulative_emissions) %>%
    mutate(c_emissions_fraction = source_quantity / cumulative_emissions) %>% 
    group_by(year, ssp) %>%
    mutate(sdev = sd(c_emissions_fraction),
           med = median(c_emissions_fraction),
           minimum = min(c_emissions_fraction),
           maximum = max(c_emissions_fraction),
           def = "Carbon tracking")
}

# Compute cumulative emissions over time
emissions <- trk_output_final %>% calc_emissions() 

# Compute the fraction of fossil fuel emissions residing in the atmosphere
emissions_fraction <- calc_emissions_fraction(trk_output_final, emissions)

rm(trk_output_final)

# Classical definition of AF - find summary stats
classic_af <- af_results %>%
  group_by(group, year, ssp) %>%
  mutate(sdev = sd(af, na.rm = TRUE),
         minimum = min(af),
         maximum = max(af),
         med = median(af),
         def = "Airborne fraction") %>% 
  filter(window_size == 1,
         run_number %in% emissions_fraction$run_number)

# Isolate relevant parameters and plot
af_plot_data <- bind_rows(emissions_fraction, classic_af) %>%
  select(run_number, year, ssp, minimum, maximum, med, sdev, def) %>%
  filter(year %in% 1910:2200) # past 2200, data gets weird

af_plot <- ggplot(af_plot_data, aes(x = year, fill = def)) +
  geom_ribbon(aes(ymin = minimum, ymax = maximum), alpha = 0.2) +
  geom_ribbon(aes(ymin = med - sdev, ymax = med + sdev), alpha = 0.4) +
  geom_line(aes(y = med, color = def)) +
  scale_color_manual(breaks = c("Airborne fraction", "Carbon tracking"),
                     values = c("#bf812d", "#35978f")) +
  scale_fill_manual(breaks = c("Airborne fraction", "Carbon tracking"),
                    values = c("#bf812d", "#35978f")) +
  facet_wrap(~ssp, scales = "free_y", 
             labeller = labeller(ssp = ssp_labels),
             ncol = 2) +
  labs(x = "Year",
       y = expression(Atmospheric~anthropogenic~CO[2]),
       color = "Computation",
       fill = "Computation") +
  theme(axis.text.x = element_text(angle = 90)) + 
  scale_x_continuous(expand = c(0, 0)) +
  ylim(-0.5, 1.5) ; af_plot

# Compute average airborne fraction across all runs
average_af <- classic_af %>%
  ungroup() %>%
  filter(year %in% 1960:2020) %>%
  summarize(avg = mean(af),
            .groups = "drop")

# Write output file
write.csv(average_af, "../output/average_af.csv")

ggsave(plot = af_plot, "../figures/figure7.png", height = 6, width = 9, units = "in")

```

### Airborne fraction trend

``` {r af-trend, message = FALSE, fig.cap = " The spread of airborne fraction decadal trend values from 1960-2020 across all runs and scenarios. See the table below for summary statistics."}
# This chunk calculates airborne fraction trend from 1960-2020 by decade, 
# as in van Marle et al. (2022)

# Extract results from above
af <- af_results %>%
  group_by(group, year, ssp) %>%
  mutate(def = "Airborne fraction") %>% 
  filter(window_size == 1,
         run_number %in% emissions_fraction$run_number,
         year %in% 1910:2200) %>%
  select(run_number, year, ssp, af, def)

# Calculate AF trend over time from 1960-2020
af_trend <- af %>%
  filter(year %in% 1960:2020) %>%
  group_by(run_number, def) %>%
  do(mod = lm(af ~ year, data = .)) %>%
  summarize(run_number = run_number,
            def = def,
            broom::tidy(mod)) %>%
  # Extract only the slope, not the intercept term
  filter(term == "year") %>%
  # Multiply annual trend by 10 for decadal average
  mutate(decade = estimate * 10,
         error = std.error * 10)

# Organize data, make table and graph
a <- af_trend %>%
  arrange(def) %>%
  select(run_number, 
         computation = def, 
         af_trend = decade, error, p.value) 

# Plot histogram of trend values
af_trend_plot <- a %>% 
  ggplot(aes(af_trend)) + 
  geom_histogram(bins = 40, fill = "#01665e") +
  geom_vline(xintercept = 0) +
  labs(x = expression(Airborne~fraction~trend~(decade^-1)),
       y = "Count") +
  scale_x_continuous(expand = c(0, 0)) ; af_trend_plot

# Summary stats
a_stats <- a %>%
  summarize(sdev = sd(af_trend),
            minimum = min(af_trend),
            maximum = max(af_trend),
            mean = mean(af_trend),
            median = median(af_trend))

a_stats %>% knitr::kable(digits = 4)

# Write output file
write.csv("../output/af_trend_summary.csv")

# What percent of runs are negative?
af_negative <- round(sum(a$af_trend < 0) / nrow(a) * 100, 1)

# Write output file
write.csv("../output/percent_runs_negative_af.csv")

ggsave(plot = af_trend_plot, "../figures/figure8.png", height = 6, width = 9, units = "in")

# Calculate land-borne and ocean-borne fractions
model_output <- fread("../output/model_output_final.csv")
  
# Ocean
ocean <- model_output %>%
  filter(variable == "ocean_uptake",
         run_number %in% unique(atm_earth_diffs$run_number),
         year %in% 1959:2015) %>%
  pivot_wider(names_from = "variable", values_from = "value") %>%
  group_by(run_number) %>%
  mutate(ocean_diff = c(NA, diff(ocean_uptake)))

oceanborne <- left_join(ocean, filter(atm_earth_diffs, year %in% 1960:2015)) %>%
  select(-atmos_co2, -atm_diff) %>%
  filter(year > 1959) %>%
  group_by(run_number, ssp) %>%
  summarize(ocean_uptake = ocean_uptake,
            earth_diff = earth_diff,
            ocean_frac = ocean_uptake / earth_diff,
            ocean_percent = ocean_frac * 100)

mean_ocean <- oceanborne %>%
  ungroup %>%
  summarize(avg = mean(ocean_percent))

# Land
land <- model_output %>%
  filter(variable == "NBP",
         run_number %in% unique(atm_earth_diffs$run_number),
         year %in% 1959:2015) %>%
  pivot_wider(names_from = "variable", values_from = "value") %>%
  group_by(run_number) %>%
  mutate(land_diff = c(NA, diff(NBP)))

landborne <- left_join(land, filter(atm_earth_diffs, year %in% 1960:2015)) %>%
  select(-atmos_co2, -atm_diff) %>%
  filter(year > 1959) %>%
  group_by(run_number, ssp) %>%
  summarize(NBP = NBP,
            earth_diff = earth_diff,
            land_frac = NBP / earth_diff,
            land_percent = land_frac * 100)

mean_land <- landborne %>%
  ungroup() %>%
  summarize(avg = mean(land_percent))

```

# The End

```{r info}
sessionInfo()
```
