==> workflow/04_supplementary_figures.Rmd <==
---
title: "04_supplementary_figures"
author: "Leeya Pressburger"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Supplementary and extra figures

```{r intro, warning=FALSE, message=FALSE}
# Load necessary packages, read in workflow settings
library(hector)
library(dplyr)
library(tidyr)
library(ggplot2)
library(GGally)
library(data.table)

theme_set(theme_light())

source("common_data.R")

# Read in data files
runlist <- fread("../output/runlist.csv")
model_output <- fread("../output/model_output_final.csv")
trk_output <- fread("../output/trk_output_final.csv")

```

## Realized parameter PDFs

``` {r s-params-PDFs, fig.cap = paste0(" Probability densities of each parameter (N=", SSP_runs[[MAIN_SCENARIO]], " draws).")}
# Density plots to visualize parameter distributions
# Create facet labels
param_labels <- c("Aerosol forcing", "CO2 fertilization", "ECS", "High-lat overturning",
                  "Intermed-deep exchange", "LUC emissions", "Ocean diffusivity", "Pre-industrial NPP",
                  "Q10", "Thermohaline overturning",  "Warm-intermed exchange")
names(param_labels) <- c("AERO_SCALE", "BETA",  "ECS","TU", "TID", "LUC_SCALE", "DIFFUSIVITY",
                         "NPP_FLUX0", "Q10_RH", "TT", "TWI")

runlist_plot <- runlist %>% 
  filter(ssp == MAIN_SCENARIO) %>%
  select(-c(run_number, ssp)) %>%
  pivot_longer(everything()) %>% 
  ggplot(aes(value)) +
  geom_density() + 
  facet_wrap(~name, scales = "free",
             labeller = labeller(name = param_labels)) +
  labs(x = NULL, y = NULL) ; runlist_plot
```

## Source coefficients of variability

``` {r s-coeff-var, eval = F, echo = FALSE, eval = FALSE, warning = FALSE, fig.cap = " Coefficients of variability for each source in the atmosphere over time."}
# Compute coefficient of variability for each source in the atmosphere pool
atmos_cv <- trk_output %>%
  filter(pool_name == "atmos_co2",
         ssp == MAIN_SCENARIO) %>%
  group_by(year, source_name) %>%
  mutate(sdev = sd(source_fraction),
         mean = mean(source_fraction),
         cv = sdev / mean)

# Plot the coefficient of variability for each source in atmosphere pool over time
ggplot(atmos_cv, aes(year, cv, group = source_name, color = source_name)) +
  geom_line() +
  scale_color_brewer(palette = "BrBG",
                     breaks = c("veg_c", "detritus_c", "soil_c", "earth_c",
                                "LL", "atmos_co2", "intermediate", "deep", "HL"),
                     labels = c("Vegetation", "Detritus", "Soil", 
                                "Deep ocean", "Low-lat ocean", 
                                "Anthropogenic emissions",
                                "Intermed. ocean", "High-lat ocean",
                                "Atmosphere")) +
  labs(x = "Source fraction",
       y = "Coefficient of variability",
       color = "Source parameter") +
  xlim(1800, 2300) +
  ylim(0, 0.25)

```

## Parameter distributions and source fraction controls

```{r s-parameter-space, eval = F, fig.cap = " Parameter correlations and distributions (diagonal), with color indicating how much of the atmosphere is composed of anthropogenic CO2 in the year 2100."}
# Exploring how the parameter space is linked to output
# What parameters control the percentage of earth_c in the atmosphere?
# Isolating the year 2100
pairs_labels <- c("Aerosol", "Beta", "ECS", 
                  "Keff",  "LUC", "NPP", "Q10", "TID", "TT", "TU", "TWI")
names(pairs_labels) <- c("AERO_SCALE", "BETA", "ECS", "DIFFUSIVITY", "LUC_SCALE", 
                         "NPP_FLUX0", "Q10_RH", "TID", "TT", "TU", "TWI")

ffi_atm_2100 <- trk_output %>%
  filter(pool_name == "atmos_co2",
         source_name == "earth_c",
         year == 2100) %>% 
  mutate(sf = cut(source_fraction, 3)) %>% 
  left_join(runlist, by = c("run_number", "ssp"))

n <- length(name_vector)

which_cols <- which(names(ffi_atm_2100) %in% names(name_vector))

ggpairs(ffi_atm_2100,
        columns = which_cols,
        # Color by magnitude of source_fraction - how much earth_c is in the atmosphere
        aes(color = sf), 
        diag = list(mapping = aes(alpha = 0.5)),
        upper = list(continuous = "blank"),
        legend = c(6,5),
        columnLabels = pairs_labels) + 
  scale_color_manual(values = c("#dfc27d", "#01665e", "#bf812d")) + 
  scale_fill_manual(values = c("#dfc27d", "#01665e", "#bf812d")) +
  labs(color = "Source Fraction")
```

## Does time affect destination?

```{r tracking-multiple-years, eval = F, warning = FALSE, fig.cap = paste0(" Final destination pools of anthropogenic emissions in ", MAIN_SCENARIO, " for ten 50-year periods between 1750 and 2250. Note that turning on carbon tracking at the beginning of each period does not have a dramatic effect on the destination of emissions.")}
# Prepare data: reset the core, set start dates every 50 years
tcore <- newcore(SSP_files[[MAIN_SCENARIO]])
dates <- seq(from = TRACKING_START, to = 2250, 50)
tracking_output <- list()

# Function to set a tracking date, run the core for a 50 year period, and retrieve tracking data
set_tracking <- function(c, date){
  setvar(c, NA, TRACKING_DATE(), date, "(unitless)")
  reset(c)
  run(c, (date + 50))
  tdata <- get_tracking_data(c)
  tdata$ty <- date
  tdata$ssp <- MAIN_SCENARIO
  tdata
}

# Loop through above function for desired dates
for(d in dates){
  tracking_output[[d]] <- set_tracking(tcore, d)
}

all_years_tracking <- bind_rows(tracking_output)

# Ordered pool labels for plotting
o_pool_labels <- c("Atmosphere", "Soil", "Vegetation",  "Detritus",
                   "High-lat ocean", "Low-lat ocean",
                   "Intermed. ocean", "Deep ocean")
names(o_pool_labels) <- c("atmos_co2", "soil_c", "veg_c",  "detritus_c",
                          "HL", "LL", "intermediate", "deep")

# Plot the destination of earth_c emissions
all_years_tracking %>%
  filter(source_name == "earth_c", pool_name != "earth_c") %>%
  mutate(source_quantity = pool_value * source_fraction) %>%
  group_by(year) %>%
  mutate(destination_fraction = source_quantity / sum(source_quantity)) %>% 
  ungroup() %>%
  mutate(pool_name = factor(pool_name, levels = rev(c("deep", "intermediate",
                                                      "LL", "HL", "detritus_c",
                                                      "veg_c", "soil_c", "atmos_co2")))) %>%
  filter(ty != "2250") %>%
  ggplot(aes(year, destination_fraction, fill = pool_name), group = ty) +
  geom_area() +
  facet_wrap(~ty, 
             scales = "free_x",
             nrow = 2) +
  labs(x = "Year",
       y = "Fraction of anthropogenic emissions",
       fill = "Destination") +
  scale_fill_brewer(palette = "BrBG",
                    direction = -1,
                    breaks = names(o_pool_labels),
                    labels = o_pool_labels) +
  theme(axis.text.x = element_text(angle = 90))

```
